requirements.txt
streamlit==1.10.0
opencv-python-headless==4.5.5.64
pyzbar==0.1.8
PyGithub==1.55
pandas==1.3.5
numpy==1.21.5
